requirements.txt
Django
djangorestframework
Pillow
channels
gunicorn
redis
python-jose[cryptography]
python-dotenv
django-imagekit
daphne
channels-redis
Twisted[http2]
orjson